obs 29 1 3 merge (#555)

+++ plugins/win-capture/graphics-hook/d3d8-capture.cpp
@@ -35,7 +35,7 @@ static d3d8_data data = {};
 
 static DXGI_FORMAT d3d8_to_dxgi_format(D3DFORMAT format)
 {
-	switch ((unsigned long)format) {
+	switch (format) {
 	case D3DFMT_X1R5G5B5:
 	case D3DFMT_A1R5G5B5:
 		return DXGI_FORMAT_B5G5R5A1_UNORM;
